check internet and keys

+++ src/screens/UploadData/index.jsx
@@ -3,6 +3,8 @@ import { View, Text } from "react-native";
 import { useNavigation } from "@react-navigation/native";
 import { RectButton } from "react-native-gesture-handler";
 
+import * as Network from "expo-network";
+
 import axios from "axios";
 
 import AsyncStorage from "@react-native-community/async-storage";
@@ -15,17 +17,26 @@ const UploadData = (props) => {
   const { navigate } = useNavigation();
 
   const handleUploadData = async () => {
-    const keys = await AsyncStorage.getAllKeys();
-    const result = await AsyncStorage.multiGet(keys);
-    await AsyncStorage.clear();
-    result.map((item) => {
-      object[`${item[0]}`] = item[1];
-    });
-    let res = await Api.postAnswers(object);
-    if (res.error == "") {
-      console.log("deu certo");
+    const { isInternetReachable } = await Network.getNetworkStateAsync();
+    if (isInternetReachable) {
+      const keys = await AsyncStorage.getAllKeys(); q
+      if (keys.length != 0) {
+        const result = await AsyncStorage.multiGet(keys);
+        await AsyncStorage.clear();
+        result.map((item) => {
+          object[`${item[0]}`] = item[1];
+        });
+        let res = await Api.postAnswers(object);
+        if (res.error == "") {
+          console.log("deu certo");
+        } else {
+          console.warn(res.error);
+        }
+      } else {
+        alert("Você precisa responder pelo menos uma vez o questionário");
+      }
     } else {
-      console.warn(res.error);
+      alert("Você está sem conexão com a internet");
     }
   };
